pyproject.toml
[project]
name = "python-sendparcel-inpost"
version = "0.1.0"
description = "InPost ShipX provider for python-sendparcel."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
authors = [{ name = "Dominik Kozaczko", email = "dominik@kozaczko.info" }]
keywords = ["shipping", "parcel", "inpost", "shipx", "sendparcel"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = [
    "python-sendparcel>=0.1.0",
    "httpx>=0.27.0",
    "anyio>=4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0",
    "respx>=0.22.0",
    "ruff>=0.9.0",
]

[project.entry-points."sendparcel.providers"]
inpost_locker = "sendparcel_inpost.providers.locker:InPostLockerProvider"
inpost_courier = "sendparcel_inpost.providers.courier:InPostCourierProvider"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sendparcel_inpost"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py312"
line-length = 80
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "N", "UP", "B", "A", "SIM", "RUF"]
ignore = ["N818", "B027"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["RUF012"]

[tool.coverage.run]
branch = true
source = ["sendparcel_inpost"]

[tool.coverage.report]
show_missing = true
